Externalize IP instead of entire URL for keycloak
This is a hack. Since frontend requests the token from keycloak using an
IP address, the issuer in the token will contain an IP address.
The AccessTokenVerifier on the keycloak adapter of every spring application
verifies the issuer present in the token with the value configured in
MultitenantConfigResolver. For these two to match, the IP is set from the
environment variable which in turn is used to configure the AdapterConfig.